+++ libs/menu-matriarch/data-access/src/lib/dtos/tag-dto.spec.ts
@@ -1,0 +1,34 @@
+import { createTagDto } from './tag-dto';
+
+describe('createTagDto', () => {
+  it('creates default tag when no arguments are passed in', () => {
+    const tag = createTagDto({});
+    expect(tag).toEqual({
+      id: '',
+      uid: '',
+      name: '',
+      color: '',
+      mealIds: [],
+      dishIds: [],
+    });
+  });
+
+  it('creates tag when arguments are passed in', () => {
+    const tag = createTagDto({
+      id: '1',
+      uid: '7R',
+      name: 'Pescatarian',
+      color: 'cadetblue',
+      dishIds: ['8', '9'],
+      mealIds: ['10'],
+    });
+    expect(tag).toEqual({
+      id: '1',
+      uid: '7R',
+      name: 'Pescatarian',
+      color: 'cadetblue',
+      dishIds: ['8', '9'],
+      mealIds: ['10'],
+    });
+  });
+});
