+++ libs/menu-matriarch/data-access/src/lib/dtos/tag-dto.ts
@@ -1,3 +1,21 @@
 import { Tag } from '@atocha/menu-matriarch/util';
 
 export type TagDto = Tag;
+
+export function createTagDto({
+  id,
+  uid,
+  name,
+  color,
+  mealIds,
+  dishIds,
+}: Partial<TagDto>): TagDto {
+  return {
+    id: id ?? '',
+    uid: uid ?? '',
+    name: name ?? '',
+    color: color ?? '',
+    mealIds: mealIds ?? [],
+    dishIds: dishIds ?? [],
+  };
+}
