+++ libs/menu-matriarch/data-access/src/lib/dtos/user-dto.ts
@@ -1,3 +1,23 @@
 import { User } from '@atocha/menu-matriarch/util';
 
 export type UserDto = User;
+
+export function createUserDto({
+  uid,
+  name,
+  email,
+  preferences,
+}: Partial<UserDto>): UserDto {
+  return {
+    uid: uid ?? '',
+    name: name ?? '',
+    email: email ?? '',
+    preferences: {
+      darkMode: preferences?.darkMode ?? false,
+      dayNameDisplay: preferences?.dayNameDisplay ?? 'full',
+      defaultMenuStartDay: preferences?.defaultMenuStartDay ?? 'Monday',
+      emptyMealText: preferences?.emptyMealText ?? 'undecided',
+      mealOrientation: preferences?.mealOrientation ?? 'horizontal',
+    },
+  };
+}
