+++ libs/menu-matriarch/data-access/src/lib/internal/dish-data.service.ts
@@ -5,8 +5,7 @@ import { map } from 'rxjs/operators';
 import { DataService } from '@atocha/core/data-access';
 import { lower, sort } from '@atocha/core/util';
 import { Dish } from '@atocha/menu-matriarch/util';
-import { DishDto } from '../dtos/dish-dto';
-import { createDishDto } from '../dtos/create-dtos';
+import { DishDto, createDishDto } from '../dtos/dish-dto';
 import { BatchService } from './batch.service';
 import { Endpoint } from './endpoint.enum';
