+++ libs/menu-matriarch/data-access/src/lib/internal/meal-data.service.ts
@@ -5,8 +5,7 @@ import { map } from 'rxjs/operators';
 import { DataService } from '@atocha/core/data-access';
 import { lower, sort } from '@atocha/core/util';
 import { Meal } from '@atocha/menu-matriarch/util';
-import { createMealDto } from '../dtos/create-dtos';
-import { MealDto } from '../dtos/meal-dto';
+import { MealDto, createMealDto } from '../dtos/meal-dto';
 import { BatchService } from './batch.service';
 import { Endpoint } from './endpoint.enum';
