+++ libs/menu-matriarch/data-access/src/lib/internal/menu-data.service.ts
@@ -5,8 +5,7 @@ import { map } from 'rxjs/operators';
 import { DataService } from '@atocha/core/data-access';
 import { flattenValues, lower, sort } from '@atocha/core/util';
 import { Day, Menu } from '@atocha/menu-matriarch/util';
-import { createMenuDto } from '../dtos/create-dtos';
-import { MenuDto } from '../dtos/menu-dto';
+import { MenuDto, createMenuDto } from '../dtos/menu-dto';
 import { BatchService } from './batch.service';
 import { Endpoint } from './endpoint.enum';
