+++ libs/menu-matriarch/data-access/src/lib/internal/tag-data.service.ts
@@ -5,8 +5,7 @@ import { map } from 'rxjs/operators';
 import { DataService } from '@atocha/core/data-access';
 import { lower, sort } from '@atocha/core/util';
 import { Tag } from '@atocha/menu-matriarch/util';
-import { createTagDto } from '../dtos/create-dtos';
-import { TagDto } from '../dtos/tag-dto';
+import { TagDto, createTagDto } from '../dtos/tag-dto';
 import { BatchService } from './batch.service';
 import { Endpoint } from './endpoint.enum';
