+++ libs/menu-matriarch/data-access/src/lib/seed-data.service.ts
@@ -2,14 +2,12 @@ import { Injectable } from '@angular/core';
 
 import { DataService } from '@atocha/core/data-access';
 import { BatchService } from './internal/batch.service';
-import {
-  createUserDto,
-  createMenuDto,
-  createMealDto,
-  createDishDto,
-  createTagDto,
-} from './dtos/create-dtos';
 import { Endpoint } from './internal/endpoint.enum';
+import { createDishDto } from './dtos/dish-dto';
+import { createMealDto } from './dtos/meal-dto';
+import { createMenuDto } from './dtos/menu-dto';
+import { createTagDto } from './dtos/tag-dto';
+import { createUserDto } from './dtos/user-dto';
 
 @Injectable({
   providedIn: 'root',
